Consolidate NodeJS adapter and its Config into one file
With the Adapters.Config macro gone, there's no longer a reason for
the NodeJS adapter's config struct to live in a separate module/file
— it has one consumer, is ~25 lines, and `@moduledoc false`. Move the
struct to a nested `Config` submodule inside `Inertia.SSR.Adapters.NodeJS`
(preserving the existing module path) and inline the `build/1` factory
into `init/1`. The `node_js/` directory disappears.
